+++ packages/flutter_tools/lib/src/commands/migrate_status.dart
@@ -6,7 +6,6 @@ import 'package:meta/meta.dart';
 
 import '../base/file_system.dart';
 import '../base/logger.dart';
-import '../base/terminal.dart';
 import '../globals.dart' as globals;
 import '../project.dart';
 import '../runner/flutter_command.dart';
@@ -62,48 +61,4 @@ class MigrateStatusCommand extends FlutterCommand {
 
     return const FlutterCommandResult(ExitStatus.success);
   }
-}
-
-/// Returns true if the migration working directory has all conflicts resolved and prints the migration status.
-bool checkAndPrintMigrateStatus(MigrateManifest manifest, Directory workingDir, {bool warnConflict = false}) {
-  List<String> remainingConflicts = <String>[];
-  List<String> mergedFiles = <String>[];
-  for (String localPath in manifest.conflictFiles) {
-    if (!MigrateUtils.conflictsResolved(workingDir.childFile(localPath).readAsStringSync())) {
-      remainingConflicts.add(localPath);
-    } else {
-      mergedFiles.add(localPath);
-    }
-  }
-  mergedFiles.addAll(manifest.mergedFiles);
-  if (manifest.addedFiles.isNotEmpty) {
-    globals.printStatus('Added files:');
-    for (String localPath in manifest.addedFiles) {
-      globals.printStatus('  - $localPath');
-    }
-  }
-  if (manifest.deletedFiles.isNotEmpty) {
-    globals.printStatus('Deleted files:');
-    for (String localPath in manifest.deletedFiles) {
-      globals.printStatus('  - $localPath');
-    }
-  }
-  if (mergedFiles.isNotEmpty) {
-    globals.printStatus('Modified files:');
-    for (String localPath in mergedFiles) {
-      globals.printStatus('  - $localPath');
-    }
-  }
-  if (remainingConflicts.isNotEmpty) {
-    if (warnConflict) {
-      globals.printWarning('Unable to apply migration. The following files in the migration working directory still have unresolved conflicts:');
-    } else {
-      globals.printStatus('Merge conflicted files:');
-    }
-    for (String localPath in remainingConflicts) {
-      globals.printStatus('  - $localPath', color: TerminalColor.red);
-    }
-    return false;
-  }
-  return true;
 }
