+++ packages/flutter_tools/lib/src/migrate/migrate_manifest.dart
@@ -8,6 +8,7 @@ import 'package:yaml/yaml.dart';
 import '../base/file_system.dart';
 import '../base/common.dart';
 import '../base/logger.dart';
+import '../base/terminal.dart';
 import '../globals.dart' as globals;
 import '../migrate/migrate_compute.dart';
 import '../migrate/migrate_config.dart';
@@ -111,7 +112,6 @@ class MigrateManifest {
     String newFileManifestContents = '';
     for (String localPath in addedFiles) {
       newFileManifestContents += '  - $localPath\n';
-      print('  Wrote Additional file $localPath');
     }
 
     String deletedFileManifestContents = '';
@@ -124,4 +124,48 @@ class MigrateManifest {
     migrateManifest.createSync(recursive: true);
     migrateManifest.writeAsStringSync(migrateManifestContents, flush: true);
   }
+}
+
+/// Returns true if the migration working directory has all conflicts resolved and prints the migration status.
+bool checkAndPrintMigrateStatus(MigrateManifest manifest, Directory workingDir, {bool warnConflict = false}) {
+  List<String> remainingConflicts = <String>[];
+  List<String> mergedFiles = <String>[];
+  for (String localPath in manifest.conflictFiles) {
+    if (!MigrateUtils.conflictsResolved(workingDir.childFile(localPath).readAsStringSync())) {
+      remainingConflicts.add(localPath);
+    } else {
+      mergedFiles.add(localPath);
+    }
+  }
+  mergedFiles.addAll(manifest.mergedFiles);
+  if (manifest.addedFiles.isNotEmpty) {
+    globals.printStatus('Added files:');
+    for (String localPath in manifest.addedFiles) {
+      globals.printStatus('  - $localPath');
+    }
+  }
+  if (manifest.deletedFiles.isNotEmpty) {
+    globals.printStatus('Deleted files:');
+    for (String localPath in manifest.deletedFiles) {
+      globals.printStatus('  - $localPath');
+    }
+  }
+  if (mergedFiles.isNotEmpty) {
+    globals.printStatus('Modified files:');
+    for (String localPath in mergedFiles) {
+      globals.printStatus('  - $localPath');
+    }
+  }
+  if (remainingConflicts.isNotEmpty) {
+    if (warnConflict) {
+      globals.printWarning('Unable to apply migration. The following files in the migration working directory still have unresolved conflicts:');
+    } else {
+      globals.printStatus('Merge conflicted files:');
+    }
+    for (String localPath in remainingConflicts) {
+      globals.printStatus('  - $localPath', color: TerminalColor.red);
+    }
+    return false;
+  }
+  return true;
 }
